Complete refactor, each DataType now has its own file
