+++ api.go
@@ -17,9 +17,16 @@ const (
 	BaseAPI = "https://api.genshin.dev/"
 )
 
-// GetCustomBody : Function to get data from API
-// ext: Slice of URL arguments for API.
-// Function will concatenate with backslashes.
+// APIError : Struct to check for API errors.
+// This struct is used in the GetCustomBody API function.
+type APIError struct {
+	Error           *string  `json:"error"`                     // If API encounters an error, Error will not be nil.
+	AvailableTypes  []string `json:"availableTypes,omitempty"`  // For wrong data type
+	AvailableImages []string `json:"availableImages,omitempty"` // For wrong image type
+}
+
+// GetCustomBody : Function to get data from API. Provide the function with a slice of URL-part strings.
+// The function will concatenate the strings with backslashes so you do not need to do so.
 func GetCustomBody(ext ...string) ([]byte, error) {
 	reqUrl := fmt.Sprintf("%s%s", BaseAPI, strings.Join(ext, "/"))
 	resp, err := http.Get(reqUrl)
@@ -39,7 +46,7 @@ func GetCustomBody(ext ...string) ([]byte, error) {
 		return nil, err
 	}
 
-	// Check error
+	// Check error. If there is an error, then the Error field will not be empty.
 	var apierror APIError
 	err = json.Unmarshal(byteArr, &apierror)
 	if apierror.Error != nil {
